Change KarmaVoteStripe.tsx to PostInteractionStripe.tsx and add "post read" state
