+++ packages/lesswrong/lib/components.ts
@@ -1137,7 +1137,7 @@ importComponent("PostNominatedNotification", () => require('../components/review
 importComponent("SingleLineReviewsList", () => require('../components/review/SingleLineReviewsList'));
 
 importComponent("QuadraticVotingButtons", () => require('../components/review/QuadraticVotingButtons'))
-importComponent("KarmaVoteStripe", () => require('../components/review/KarmaVoteStripe'))
+importComponent("PostInteractionStripe", () => require('../components/review/PostInteractionStripe'))
 importComponent("ReviewVoteTableRow", () => require('../components/review/ReviewVoteTableRow'))
 importComponent("ReviewVotingButtons", () => require('../components/review/ReviewVotingButtons'))
